refactor(ui): split FilterSubscriptionList into table/form/actions
